refactor: 예외 팩토리를 사용하도록 변경
